+++ Leomard/Models/Leomard/LeomardExceptions.swift
@@ -14,6 +14,7 @@ enum LeomardExceptions: Error, CustomStringConvertible {
     case notLoggedIn(String)
     case unableToGetIcon(String)
     case userCancelledOperation(String)
+    case nsfwPost(String)
     
     var description: String {
         switch self {
@@ -28,6 +29,8 @@ enum LeomardExceptions: Error, CustomStringConvertible {
         case .unableToGetIcon(let message):
             return message
         case .userCancelledOperation(let message):
+            return message
+        case .nsfwPost(let message):
             return message
         }
     }