+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.6 (TBA)
+
+### Added
+
+- "Store Liked Posts" in Preferences -> Experiments
+    - With this enabled, you can see all the posts you liked in the "Liked" tab in your profile
+
+### Changes
+
+- "Two-Colum View" is now the default view
+
 ## 0.5.1 (TBA)
 
 ### Bug Fixes